add slist/forward_list
slist: optimize transfers into before and after
no run-time branches

stop at end of range not at end of list.
The STL needs before_begin() as forward_list start iterator in some cases.
gen_iters still not fully working though.

wrote a merge_sort. The fast sort from list does not work yet, but this one
is as fast as the C++ counterpart. so good enough.

+++ tests/test.h
@@ -116,7 +116,7 @@ static inline long TEST_TIME(void)
     }                                                                                                                  \
     loops:
 
-#define RECORD_WHICH covvec.vector[which]++
+#define RECORD_WHICH covvec.vector[(u16)which]++
 
 /* check if we covered all tests. If not redo the missing tests. */
 static inline int check_redo(vec_u16 *covvec, queue_int *tests, const int number_ok)
